configure: use pkg-config for FreeType 2
In package-based Linux distributions FreeType 2 headers tend to be in
/usr/include/freetype2, in the FreeBSD port they are in
/usr/local/include/freetype2 and in the NetBSD pkgsrc package they are
in /usr/pkg/include/freetype2. pkg-config is the established means of
managing this compexity for both the compiler and the linker flags.

GRASS has been using both FreeType 2 and pkg-config, but not the two
together, and its default FreeType 2 include directory has been set to
/usr/include/freetype, which belongs to FreeType 1. This way, the
configure script always required the --with-freetype-includes=...
argument.

Fix the script to handle FreeType 2 the same way it handles Cairo: allow
to override the defaults if required, but default to the flags provided
by pkg-config so things usually just work:

$ ./configure
[...]
checking whether to use FreeType... yes
checking for location of FreeType includes...
checking for ft2build.h... yes
checking for location of FreeType library...
checking for FT_Init_FreeType in -lfreetype... yes
[...]